Version 2.0.0 Update

+++ fields/acf-icon-picker-v5.php
@@ -2,15 +2,15 @@
 
 if( ! defined( 'ABSPATH' ) ) exit;
 
-if( !class_exists('acf_field_icon_picker') ) :
+if( !class_exists('nc_acf_field_icon_picker') ) :
 
-class acf_field_icon_picker extends acf_field {
+class nc_acf_field_icon_picker extends acf_field {
 
 	function __construct( $settings ) {
 
-		$this->name = 'icon-picker';
+		$this->name = 'nc-acf-icon-picker';
 
-		$this->label = __('Icon Picker', 'acf-icon-picker');
+		$this->label = __('NC Icon Picker', 'acf-icon-picker');
 
 		$this->category = 'jquery';
 
@@ -24,30 +24,71 @@ class acf_field_icon_picker extends acf_field {
 
 		$this->settings = $settings;
 
-		$this->path_suffix = apply_filters( 'acf_icon_path_suffix', 'assets/img/acf/' );
+		$this->path_suffix = apply_filters( 'nc_acf_icon_path_suffix', 'images/svg-icons' );
+		$this->parent_path_suffix = apply_filters( 'nc_acf_icon_parent_path_suffix', false );
 
-		$this->path = $this->settings['path'] . $this->path_suffix;
-
-		$this->url = $this->settings['url'] . $this->path_suffix;
-
-		$priority_dir_lookup = get_stylesheet_directory() . '/' . $this->path_suffix;
-
-		if ( file_exists( $priority_dir_lookup ) ) {
-			$this->path = $priority_dir_lookup;
-			$this->url = get_stylesheet_directory_uri() . '/' . $this->path_suffix;
+		if (! $this->parent_path_suffix ) {
+			$this->parent_path_suffix = $this->path_suffix;
 		}
+
+		$this->path = get_stylesheet_directory() . '/' . $this->path_suffix;
+
+		if ( is_dir( $this->path ) ) {
+			$this->url = get_stylesheet_directory_uri() . '/' . $this->path_suffix;
+		} else {
+			$this->path = $this->settings['path'] . $this->path_suffix;
+			$this->url = $this->settings['url'] . $this->path_suffix;
+		}
+
+		if ($this->parent_path_suffix) {
+			$this->parent_path = get_template_directory() . '/' . $this->parent_path_suffix;
+		}
+		if ( isset($this->parent_path) && is_dir( $this->parent_path ) ) {
+			$this->parent_url = get_template_directory_uri() . '/' . $this->path_suffix;
+		} else {
+			$this->parent_path = false;
+			$this->parent_url = false;
+		}
+
 
 		$this->svgs = array();
 
-		$files = array_diff(scandir($this->path), array('.', '..'));
-		foreach ($files as $file) {
-			if( pathinfo($file, PATHINFO_EXTENSION) == 'svg' ){
-				$exploded = explode('.', $file);
-				$icon = array(
-					'name' => $exploded[0],
-					'icon' => $file
-				);
-				array_push($this->svgs, $icon);
+		$is_path = is_dir($this->path);
+		
+		$files = is_dir($this->path) ? scandir($this->path) : array();
+		$files = array_map(function ($file) {
+			return $this->path . $file;
+		}, $files);
+		
+		$parent_files = is_dir($this->parent_path) ? scandir($this->parent_path) : array();
+		$parent_files = array_map(function ($file) {
+			if ( in_array ($file, array('.', '..'))) {
+				return $file;
+			}
+			return $this->parent_path . $file;
+		}, $parent_files);
+		
+		$files = array_replace($parent_files, $files);
+
+		if (count($files)) {
+			
+			$files = array_diff($files, array('.', '..'));
+			foreach ($files as $file) {
+				if( pathinfo($file, PATHINFO_EXTENSION) == 'svg' ){
+					$name = basename($file, '.svg');
+					$icon = array(
+						'name' => $name,
+						'icon' => $name . '.svg'
+					);
+
+					$icon_location = strpos($file, $this->parent_path) !== false ? 'parent' : 'current';
+
+					$icon['location'] = $icon_location;
+
+					if ($icon_location !== 'parent' || ! file_exists($this->path . $name . '.svg')) {
+						array_push($this->svgs, $icon);
+					}
+				}
 			}
 		}
 
@@ -55,16 +96,41 @@ class acf_field_icon_picker extends acf_field {
 	}
 
 	function render_field( $field ) {
-		$input_icon = $field['value'] != "" ? $field['value'] : $field['initial_value'];
-		$svg = $this->path . $input_icon . '.svg';
+		$input_value = $field['value'] != "" ? $field['value'] : $field['initial_value'];
+		$input_array = json_decode($input_value, JSON_OBJECT_AS_ARRAY);
+		$svg = array();
+		if ($input_array === NULL) {
+			if ($input_value) {
+				$input_array = array(
+					'icon' => $input_value
+				);
+			}
+		}
+		if (isset($input_array['icon'])) {
+			$svg['icon'] = basename($input_array['icon'], '.svg');
+			$svg['path'] = $this->path . $svg['icon'] . '.svg';
+			if (file_exists($svg['path'])) {
+				$svg['url'] = $this->url . $svg['icon'] . '.svg';
+				$svg['location'] = 'current';
+			}
+			elseif ( $this->parent_path ) {
+				$svg['path'] = $this->parent_path . $svg['icon'] . '.svg';
+				$svg['url'] = $this->parent_url . $svg['icon'] . '.svg';
+				$svg['location'] = 'parent';
+			}
+
+			$svg_encoded = json_encode($svg);
+		} else {
+			$svg_encoded = null;
+		}
+		
 		?>
 			<div class="acf-icon-picker">
 				<div class="acf-icon-picker__img">
 					<?php
-						if ( file_exists( $svg ) ) {
-							$svg = $this->url . $input_icon . '.svg';
+						if ( isset( $svg['path'] ) && file_exists( $svg['path'] ) ) {
 							echo '<div class="acf-icon-picker__svg">';
-						   	echo '<img src="'.$svg.'" alt=""/>';
+						   	echo '<img src="'.$svg['url'].'" alt=""/>';
 						    echo '</div>';
 						}else{
 							echo '<div class="acf-icon-picker__svg">';
@@ -72,7 +138,7 @@ class acf_field_icon_picker extends acf_field {
 						    echo '</div>';
 						}
 					?>
-					<input type="hidden" readonly name="<?php echo esc_attr($field['name']) ?>" value="<?php echo esc_attr($input_icon) ?>"/>
+					<input type="hidden" readonly name="<?php echo esc_attr($field['name']) ?>" value='<?php echo $svg_encoded ?>'/>
 				</div>
 				<?php if ( $field['required' ] == false ) { ?>
 					<span class="acf-icon-picker__remove">
@@ -93,6 +159,7 @@ class acf_field_icon_picker extends acf_field {
 
 		wp_localize_script( 'acf-input-icon-picker', 'iv', array(
 			'path' => $this->url,
+			'parent_path' => $this->parent_url,
 			'svgs' => $this->svgs,
 			'no_icons_msg' => sprintf( esc_html__('To add icons, add your svg files in the /%s folder in your theme.', 'acf-icon-picker'), $this->path_suffix)
 		) );
@@ -101,7 +168,7 @@ class acf_field_icon_picker extends acf_field {
 		wp_enqueue_style('acf-input-icon-picker');
 	}
 }
-new acf_field_icon_picker( $this->settings );
+new nc_acf_field_icon_picker( $this->settings );
 
 endif;
